Move refilter code from MonoRepoProcessor to expander.rs
MonoRepoProcessor is now simply a container of stuff useful during
processing.
